fix: Make catalog listings executable for optional and decimal params (#2198)

+++ src/Indicators/r-s/Renko/Renko.Catalog.cs
@@ -10,7 +10,7 @@ public static partial class Renko
             .WithName("Renko Chart")
             .WithId("RENKO")
             .WithCategory(Category.PriceTransform)
-            .AddParameter<double>("brickSize", "Brick Size", description: "The size of each Renko brick", isRequired: true, defaultValue: 1.0, minimum: 0.001, maximum: 1000000.0)
+            .AddParameter<decimal>("brickSize", "Brick Size", description: "The size of each Renko brick", isRequired: true, defaultValue: 1.0m, minimum: 0.001, maximum: 1000000.0)
             .AddEnumParameter<EndType>("endType", "End Type", description: "The price candle end type to use as the brick threshold", isRequired: false, defaultValue: EndType.Close)
             .AddResult("Open", "Open", ResultType.Default)
             .AddResult("High", "High", ResultType.Default)
